Update ABC_Pipeline.py
The idea is to define a function protocol to inject the functions into a @abstractmethod and @classmethod function. 

The new function will take a tuple of functions and add them into the run method to have the pipeline setup by injection, leaving room to direct inherit from the class and define regular methods.